move _create_action to test utils
