+++ posthog/hogql_queries/insights/funnels/test/test_funnel.py
@@ -6,6 +6,7 @@ from freezegun import freeze_time
 from posthog.test.base import (
     APIBaseTest,
     ClickhouseTestMixin,
+    _create_action,
     _create_event,
     _create_person,
     also_test_with_materialized_columns,
@@ -75,14 +76,6 @@ class PseudoFunnelActors:
             [{"id": x[0]} for x in actors],
             None,
         )
-
-
-def _create_action(**kwargs):
-    team = kwargs.pop("team")
-    name = kwargs.pop("name")
-    properties = kwargs.pop("properties", {})
-    action = Action.objects.create(team=team, name=name, steps_json=[{"event": name, "properties": properties}])
-    return action
 
 
 def funnel_test_factory(event_factory, person_factory):
