+++ posthog/queries/funnels/test/test_funnel_unordered.py
@@ -3,6 +3,7 @@ from datetime import datetime
 from posthog.test.base import (
     APIBaseTest,
     ClickhouseTestMixin,
+    _create_action,
     _create_event,
     _create_person,
     snapshot_clickhouse_queries,
@@ -11,7 +12,6 @@ from posthog.test.base import (
 from rest_framework.exceptions import ValidationError
 
 from posthog.constants import INSIGHT_FUNNELS
-from posthog.models.action import Action
 from posthog.models.filters import Filter
 from posthog.queries.funnels.funnel_unordered import ClickhouseFunnelUnordered
 from posthog.queries.funnels.funnel_unordered_persons import ClickhouseFunnelUnorderedActors
@@ -24,14 +24,6 @@ from posthog.queries.funnels.test.conversion_time_cases import funnel_conversion
 from posthog.test.test_journeys import journeys_for
 
 FORMAT_TIME = "%Y-%m-%d 00:00:00"
-
-
-def _create_action(**kwargs):
-    team = kwargs.pop("team")
-    name = kwargs.pop("name")
-    properties = kwargs.pop("properties", {})
-    action = Action.objects.create(team=team, name=name, steps_json=[{"event": name, "properties": properties}])
-    return action
 
 
 class TestFunnelUnorderedStepsBreakdown(
